[LIB UPDATE] Add preprocessor check for libnl3 >= 3.5 in HTB.
HTB makes use of the libnl3 3.5 >= APIs:
```c
rtnl_htb_set_ceil64()
rtnl_htb_set_rate64()
```

This commit enforces the that the correct version is used where this
new 64 bit API is used.

+++ src/slave/containerizer/mesos/isolators/network/port_mapping.cpp
@@ -5294,10 +5294,10 @@ string PortMappingIsolatorProcess::scripts(Info* info)
     script << "tc class add dev " << eth0 << " parent "
            << CONTAINER_TX_HTB_HANDLE << " classid "
            << CONTAINER_TX_HTB_CLASS_ID << " htb rate "
-           << static_cast<uint64_t>(info->egressConfig->rate) * 8 << "bit";
+           << info->egressConfig->rate * 8 << "bit";
     if (info->egressConfig->ceil.isSome()) {
       script << " ceil "
-             << static_cast<uint64_t>(info->egressConfig->ceil.get()) * 8
+             << info->egressConfig->ceil.get() * 8
              << "bit";
       if (info->egressConfig->burst.isSome()) {
         // Use bytes here because tc command borks at bits.
